Drawing two objects, triangle and rectangle.
Added:
ObjectManager, now gets used to store all the Objects that are drawn to the screen.

Renderer, added a Pointer to the m_RenderObjects from ObjectManager to go through the entire list of Objects.

Changed:
Renderer, CreateCommandBuffers function changed to work with multiple Objects. Not sure if this is the optimal way of doing it.