Classes for Shapes and ObjectManager intergration.
Added:
BaseRenderObject. the baseclass for all RenderObjects that will be created.

RenderShape, a simple RenderObject that asks for a pre-build Mesh.

RenderFactory, a factory design that creates a RenderShape depending on what ShapeType you want, such as a Triangle or a rectangle.

Transform, this class holds basic information for the Transform of the Object such as the Position, Rotation and scale including the needed matrices except the model matrix.

ObjectManager, added m_PointerToMeshes vector variable so that the Renderer can use it.

Changed:
ObjectManager now creates Objects after creation by asking for CreateShape()