Showing Images.
Added:
ImageHandler, handles parts of the ImageCreating process for Textures or the SwapChain.

Background.png, replaces texture.jpg as the problem with getting out of memory is fixed.

Renderer, SetupImage(), function that handles the creation of a texture.
Renderer, SetupHandlers(), needed to be called later because it needs the information that the Physical device has.

Changed:
ObjectManager, m_Textures is now a Vector instead of an array.

ShaderManager/Renderer, changed the Descriptor to work with Images.

Unlit.vert/frag shaders, now works with Textures.