Rewrote Depth/ImageHandler to work with the new VulkanDevice.
Add:
AnisotropyFilteringMax, float, used to see what the max AnsiotropyFiltering is for the ImageHandler.
p_GraphicsQueue VkQueue*, temporary value used to speed up the test if the rework is a success or not.

Changed:
DepthHandler.h/cpp, now only asks for a VulkanDevice and ImageHandler on creation.
ImageHandler.h/cpp, now only asks for a VulkanDevice on creation.

Renderer.h/cpp, some new code to work better with the new handlers.