ObjectPool tool
Add:
ObjectPool.h, an template datastructure that will be used for the PipelineData and DescriptorData to make construction be made on startup so that on run-time the creation will be faster.

PipelineData.h, PipelineData + DescriptorData, now have a uint32_t for it's ID so that I can better call this data.
PipelineData.h, PipelineData, removed the constructor for now and made r_DescriptorData into a pointer renamed to p_DescriptorData.

Note:
This is overoptimizing while the functionality isn't even there, but I made it for research purposes and just as a learning opportunity.