Resizing works, now finishing to correctly shut the application off, also some memory issues when resizing.
Bugs might still appear since I programmed for the future (Bad habbit I know.)

One of the bigger changes is the use of uint32_t ID's (similiar to OpenGL) for the creation and storing of Pipeline and DescriptorData. Might want to use this standard in the future.

Change:
ObjectPool.h, GetFilledObject(), now returns the template object and asks for a reference of the bool.

ShaderManager.h/cpp, now mainly does the creation AND storage of the Descriptors and Pipelines. Using ObjectPool instead of Vector for the DescriptorData and PipelineData for further speedup later.

RenderShape/BaseRenderObject/RenderFactory.h/cpp, now using the ID system instead of a pointer.

ObjectManager.h/cpp, doesn't store the pipeline and descriptorData anymore. This is now handles by the ShaderManager and now the objectmanager just stores the ID's.