Abstracted TextureData
Add:
TextureData.h, this will now function as the main header for the TextureData struct.
TextureData.h, added descriptorInfo so that it can get the information needed for the eventual descriptor, UpdateDescriptor() can be called per frame to keep this up to date,

Change:
RenderObjectData.h, removed TextureData because of the new Headerfile.