Abstraction of SwapChain
Add:
VulkanSwapChain.h/cpp, added CreateSwapChain() and CreateImageView() functions, moved from Renderer.h/cpp to VulkanSwapChain.h/cpp.

VulkanSwapChain.h, SwapChainSupportDetails, an inline function with no class attached to it. This way you can get the information without calling the VulkanSwapChain.

Change:
VulkanDevice.h/cpp, now holds the only way to get the QueueFamilies with the Function called FindQueueFamilies (Firstly named SetQueueFamilies). It's inline so you do not need to call a variable to access this function.

VulkanSwapChain.h./cpp, VulkanDevice is now a pointer instead of a reference so that this can be a normal variable for Renderer.

Renderer.h/cpp, m_VulkanSwapChain is now not a pointer anymore.