Camera + SceneObject abstraction
Add:
SceneObject.h/cpp, this class represents a raw object on the world that has no visual representation. This is used for camera's and later can be accessed for a gameplay engine without accessing many graphical options.

CameraObject.h/cpp, this SceneObject represents a camera that can appear in a scene. It has no visualisation so it's not a BaseRenderObject.

Change:
BaseRenderObject.h/cpp, now inherits from SceneObject and gets it's transform from this class.

SceneObject, m_ID is now a uint32_t like most ID's in this engine.