CameraController and new way of uniform buffers.
New CameraSystem in place and replaces the hardcodes stuff, not all is done and clean yet though.

Add:
CameraController.h/cpp, this class handles all the cameras on the scene. It needs some serious look at with the communication between the Renderer for the ViewProjection data.

Renderer.h/cpp, now has a pointer to the active Camera, this communication is still bad and currently hardcoded on startup of the ObjectManager.

Renderer.cpp, UpdateUniformBuffer, now doesn't calculate the camera inside the function but gets the Camera Info from the pointer stored in the class.

Change:
VulkanDevice.h.cpp & ShaderManager.h/cpp, instead of getting the buffersize for ViewProjection the function to create the buffers now asks for a VkDeviceSize in it's parameters. This way the function is more abstract and universal.

CameraObject.h/cpp, some changes to accommodate the new way the camera should work.

Removed:
BufferData.h, ViewProjection, removed it and moved the alignas to the CameraObject.h ViewProjection struct.