SceneManager and Scenes.
Add:
SceneManager.h/cpp, a class that is used to manage scenes. This also replaces most of the code in main.cpp for now.
Scene.h/cpp, a class that holds information that a scene needs. A name, what 3d objects it holds and the special sceneData.

SceneData.h, this struct used to send specific graphical data for a scene like sunlight and fog color.

RenderFactory.h/cpp, CheckRenderObject(), this object will see if the renderobject was already created. If not it will call CreateRenderObject().
RenderFactory.h/cpp, CreateRenderObject(), this private function will construct RenderObjectData.

Changed:
RenderFactory.h/cpp, renamed all ...SceneObjects functions to ...RenderObjects.
RenderFactory.h/cpp, CreateRenderObject(), now also asks for a bool that decides if the object is new to the scene.

ObjectManager.h/cpp, CreateShape(), renamed to AddRenderObject() and now only asks for a BaseRenderObject*, it gets the data from SceneManager.

Removed:
ContextData.h, isn't needed anymore and forgot to remove this some commits ago.