Rectangle added to scene, multiple textures different Descriptorsets
Disabled:
Disabled the camera switch for now, was annoying to look at.

Add:
ObjectManager.h, added pip_Triangle and des_Triangle for the pipeline and descriptorset.
ObjectManager.cpp, SetupStartObjects(), now sets up the texture for the rectangle and creates it's descriptors.

Change:
Renderer.cpp, DrawObjects(), when the pipeline has changed now create a new pipeline variable instead of reusing a variable, made the textures bug out and be set as one.